move dump_execution to its own mod
